backend/requirements.txt
alembic>=1.18.4
asyncpg>=0.31.0
python-dotenv>=0.9.9
fastapi[standard]>=0.135.1
psycopg2>=2.9.11
psycopg[binary]>=3.3.3
pydantic-settings>=2.13.1
sqlalchemy>=2.0.48
structlog>=25.5.0
